Implement `refine` for `ViewRefine`
